+++ index.js
@@ -1,50 +1,56 @@
 'use strict';
 
-var child_process = require('child_process');
-var fs = require('fs');
-var path = require('path');
+import child_process from 'child_process';
+import {statSync} from 'fs';
+import path from 'path';
+import Mocha from 'mocha';
 
-var _ = require('lodash');
-var Mocha = require('mocha');
-var statSync = require('fs').statSync;
-var Reporter = require('./lib/reporter');
-var watcher = require('./lib/watcher');
+import Reporter from './lib/reporter';
+import {
+    addTest,
+    runTests,
+    setOptions as setWatcherOptions
+} from './lib/watcher';
 
 // files lookup in mocha is complex, so it's better to just run original code
-var mochaLookupFiles = require('mocha/lib/utils').lookupFiles;
+import {lookupFiles as mochaLookupFiles} from 'mocha/lib/utils';
 
-module.exports = function MochaParallelTests(options) {
+export default function MochaParallelTests(options) {
     process.setMaxListeners(0);
 
-    var extensions = ['js'];
-    (options.compilers || []).forEach(function (compiler) {
-        var compiler = c.split(':');
-        var ext = compiler[0];
+    if (typeof options.compilers === 'string') {
+        options.compilers = [options.compilers];
+    }
+
+    const extensions = ['js'];
+    (options.compilers || []).forEach(compiler => {
+        const compilerName = compiler.split(':');
+        const ext = compilerName[0];
 
         extensions.push(ext);
     });
 
     // get test files with original mocha utils.lookupFiles() function
-    var files = [];
-    options._.forEach(function (testPath) {
+    let files = [];
+    options._.forEach(testPath => {
         files = files.concat(mochaLookupFiles(testPath, extensions, options.recursive));
     });
 
     // watcher monitors running files
-    watcher.setOptions({
+    setWatcherOptions({
         maxParallelTests: options.maxParallel,
         retryCount: options.retry
     });
 
-    files.forEach(function (file, index) {
-        var testOptions = _.assign({}, options, {
+    files.forEach(file => {
+        var testOptions = Object.assign({}, options, {
             reporterName: options.R || options.reporter,
             reporter: Reporter,
             testsLength: files.length
         });
 
-        watcher.addTest(path.resolve(file), testOptions);
+        addTest(path.resolve(file), testOptions);
     });
 
-    watcher.runTests();
-};
+    runTests();
+}