+++ test/index.sh
@@ -58,6 +58,9 @@ echo $?
 echo 'TESTCASE: timeouts exit code'
 test test/timeouts-exit-code/index.sh
 echo $?
+echo 'TESTCASE: js compilers support'
+test test/js-compilers/index.sh
+echo $?
 
 echo "Passes: $PASSES Failes: $FAILES"