+++ test/missing-test/index.js
@@ -3,7 +3,7 @@
 var assert = require('assert');
 var path = require('path');
 var exec = require('child_process').exec;
-var libExecutable = path.resolve(__dirname, '../../bin/mocha-parallel-tests');
+var libExecutable = path.resolve(__dirname, '../../dist/bin/mocha-parallel-tests');
 
 exec(libExecutable + ' -R doc --timeout 60000 --slow 30000 test/missing-test/missing.js', {
     cwd: path.resolve(__dirname, '../../')