+++ test/parallel-order/index.js
@@ -3,7 +3,7 @@
 var assert = require('assert');
 var path = require('path');
 var exec = require('child_process').exec;
-var libExecutable = path.resolve(__dirname, '../../bin/mocha-parallel-tests');
+var libExecutable = path.resolve(__dirname, '../../dist/bin/mocha-parallel-tests');
 var start = Date.now();
 
 exec(libExecutable + ' -R json --timeout 60000 --slow 30000 test/parallel-order/tests', {